requirements.txt
discord.py==2.1.0
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
tinydb==4.7.0
